asl_pipeline/backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
deep-translator
python-multipart
openai-whisper
gtts